+++ pennylane_lightning/lightning_tensor/lightning_tensor.py
@@ -200,10 +200,9 @@ class LightningTensor(Device):
             be described but the larger the memory requirements. Note that GPUs are ill-suited (i.e. less
             competitive compared with CPUs) for simulating circuits with low bond dimensions and/or circuit
             layers with a single or few gates because the arithmetic intensity is lower.
-        cutoff (float): The threshold used to truncate the singular values of the MPS tensors. Default is 0
-        cutoff_mode (str): Singular value truncation mode. Options: ["rel", "abs"]. Default is "abs".
+        cutoff (float): The threshold used to truncate the singular values of the MPS tensors. The default is 0.
+        cutoff_mode (str): Singular value truncation mode. The options are ``"rel"`` and ``"abs"``. The default is ``"abs"``.
         backend (str): Supported backend. Currently, only ``cutensornet`` is supported.
-        method (str): Supported method. Currently, only ``mps`` is supported.
 
     **Example**