+++ .github/CHANGELOG.md
@@ -84,6 +84,9 @@
 
 ### Documentation
 
+* Add installation instructions and documentation for `lightning.tensor`.
+  [(#756)](https://github.com/PennyLaneAI/pennylane-lightning/pull/756)
+
 ### Bug fixes
 
 * Fix random CI failures for `lightning.tensor` python unit tests and ignore `lightning_tensor` paths.